Refactor pytest entrypoint to include missing integration tests
